Cleanup file structure.
