+++ http/Response.cpp
@@ -1,6 +1,7 @@
 #include "Response.h"
 #include "HTTPContext.h"
 #include "HTTPStatusCodes.h"
+#include "Cookie.h"
 
 
 Response::Response(HTTPContext* httpContext) : httpContext(httpContext) {
@@ -58,51 +59,3 @@ void Response::end() const {
     this->httpContext->end();
 }
 
-void Cookie::Domain(const std::string &domain) {
-    this->options.insert({ "Domain", domain});
-}
-
-void Cookie::Path(const std::string &path) {
-    this->options.insert({ "Path", path});
-}
-
-void Cookie::MaxAge(const std::string &maxAge) {
-    this->options.insert({ "Max-Age", maxAge});
-}
-
-void Cookie::Expires(const std::string &expires) {
-    this->options.insert({ "Expires", expires});
-}
-
-void Cookie::MakeEssential() {
-    this->options.insert({ "IsEssential", "true"});
-}
-
-void Cookie::MakeSecure() {
-    this->options.insert({ "Secure", "true"});
-}
-
-void Cookie::MakeHttpOnly() {
-    this->options.insert({ "HttpOnly", "true"});
-}
-
-void Cookie::SetSameSite(const bool strict) {
-    this->options.insert( { "SameSite", strict ? "Strict" : "None" });
-}
-
-void Cookie::reset() {
-    this->value = "";
-    this->options.clear();
-}
-
-const std::string &Cookie::getName() const {
-    return name;
-}
-
-const std::string &Cookie::getValue() const {
-    return value;
-}
-
-const std::map<std::string, std::string> &Cookie::getOptions() const {
-    return options;
-}
